### 2024-04-22 #174 @jinningwang
Submission for AMS
Submitting Author: Jinning Wang (@jinningwang)
All current maintainers: @jinningwang
Package Name: AMS
One-Line Description of Package: Power system dispatch modeling and dispatch-dynamic co-simulation.
Repository Link:  https://github.com/CURENT/ams
Version submitted:   v0.9.6
Version accepted: TBD

---

As part of [CURENT Large-scale Testbed](https://github.com/CURENT) platform, AMS serves as power system production cost modeling. Our framework offers a modularized approach that seamlessly incorporates dynamics, enhancing traditional dispatch modeling methods. We create a versatile solution that bridges the gap between device-level and system-level models. The tool is developed to be extensible, scalable, compatible, and interoperable.

> [^1]: Please fill out a pre-submission inquiry before submitting a data visualization package.

Power system researchers and engineers.
Power system steady state modeling and analysis.

There are some Python packages cover part of our functions: [PYPOWER](https://github.com/rwl/PYPOWER), [pandapower](https://github.com/e2nIEE/pandapower), and [PyPSA](https://github.com/PyPSA/PyPSA).

Compared to existing tools that focus on fixed power system optimization problems, our package AMS enables customizing formulations thus enable rapid prototyping for renewables integration.

Additionally, with the built-in interface with dynamic simulator ANDES, AMS allows native interoperation between dynamics and dispatch, which significantly relieves the researchers manual efforts when conducting power system simulations.

https://github.com/pyOpenSci/software-submission/issues/169
@Batalex
