### 2024-04-24 #174 @Batalex
Hi there! Thank you for submitting your package for pyOpenSci review. Below are the basic checks that your package needs to pass to begin our review. If some of these are missing, we will ask you to work on them before the review process begins.

Please check our [Python packaging guide](https://www.pyopensci.org/python-package-guide) for more information on the elements below.

  **Just so you know, I was unable to import the package on OS X with the following error: `symbol not found in flat namespace '_klu_l_analyze'`. I'll report it properly later.**

---
---
