### 2024-04-25 #174 @jinningwang
Hi @Batalex,

Thanks for your response.

> Just so you know, I was unable to import the package on OS X with the following error: symbol not found in flat namespace '_klu_l_analyze'. I'll report it properly later.

We run into it before and have addressed it, https://github.com/CURENT/andes/issues/508

Another possible solution is to add the used environment to path:

```bash
export PATH="/path/to/conda/envs/ams/bin:$PATH"
```
