### 2024-05-03 22639ed
[WIP] Add EV aggregator model
