+++ ams/extension/eva.py
@@ -1,0 +1,11 @@
+"""
+EV Aggregator.
+"""
+
+class mcs():
+    """
+    Class for EV aggregation Monte Carlo simulation.
+    """
+
+    def __init__(self):
+        pass
