### 2024-05-03 43631cd
Draft EV scheduling model
