+++ ams/models/__init__.py
@@ -11,7 +11,7 @@ ams_file_classes = list([
     ('static', ['PQ', 'Slack', 'PV']),
     ('shunt', ['Shunt']),
     ('line', ['Line']),
-    ('distributed', ['PVD1', 'ESD1']),
+    ('distributed', ['PVD1', 'ESD1', 'EV1', 'EV2']),
     ('renewable', ['REGCA1', 'REGCV1', 'REGCV2']),
     ('area', ['Area']),
     ('region', ['Region']),
