+++ ams/models/distributed/__init__.py
@@ -1,2 +1,3 @@
 from ams.models.distributed.pvd1 import PVD1  # NOQA
 from ams.models.distributed.esd1 import ESD1  # NOQA
+from ams.models.distributed.ev import EV1, EV2  # NOQA
