+++ ams/extension/eva.py
@@ -2,6 +2,7 @@
 EV Aggregator.
 """
 
+
 class mcs():
     """
     Class for EV aggregation Monte Carlo simulation.
