### 2024-05-04 cb6ec78
Add parameter N in EV model
