### 2024-05-04 b1c4ffb
[WIP] EVA, normal distribution parameter initialization
