+++ ams/extension/eva.py
@@ -2,11 +2,203 @@
 EV Aggregator.
 """
 
+from collections import OrderedDict
 
-class mcs():
+import scipy.stats as stats
+
+from andes.core import Config
+from andes.core.param import NumParam
+from andes.core.model import ModelData
+from andes.shared import pd
+
+from ams.core.model import Model
+
+
+class EVA(ModelData, Model):
     """
-    Class for EV aggregation Monte Carlo simulation.
+    State space modeling based EV aggregation model.
+
+    In the EVA, each single EV is recorded as a device with its own parameters.
+    The parameters are generated from given statistical distributions.
+
+    Reference:
+    [1] J. Wang et al., "Electric Vehicles Charging Time Constrained Deliverable Provision of Secondary
+    Frequency Regulation," in IEEE Transactions on Smart Grid, doi: 10.1109/TSG.2024.3356948.
+    [2] M. Wang, Y. Mu, Q. Shi, H. Jia and F. Li, "Electric Vehicle Aggregator Modeling and Control for
+    Frequency Regulation Considering Progressive State Recovery," in IEEE Transactions on Smart Grid,
+    vol. 11, no. 5, pp. 4176-4189, Sept. 2020, doi: 10.1109/TSG.2020.2981843.
     """
 
-    def __init__(self):
-        pass
+    def __init__(self, N=10000, Ns=20, Tagc=4, SOCf=0.2, r=0.5,
+                 seed=None,):
+        """
+        Initialize the EV aggregation model.
+
+        Parameters
+        ----------
+        N: int, optional
+            Number of related EVs, default is 10000.
+        Ns : int, optional
+            Number of SOC intervals, default is 20.
+        Tagc : int, optional
+            AGC time intervals in seconds, default is 4.
+        SOCf : float, optional
+            Force charge SOC level between 0 and 1, default is 0.2.
+        r : float, optional
+            Ratio of time range 1 to time range 2 between 0 and 1, default is 0.5.
+        """
+        # inherit attributes and methods from ANDES `ModelData` and AMS `Model`
+        ModelData.__init__(self)
+        Model.__init__(self, system=None, config=None)
+
+        # manually set config as EVA is not processed by the system
+        self.config = Config(self.__class__.__name__)
+        self.config.add(OrderedDict((('ns', Ns),
+                                     ('tagc', Tagc),
+                                     ('socf', SOCf),
+                                     ('r', r),
+                                     ('socl', 0),
+                                     ('socu', 1),
+                                     ('seed', seed),
+                                     )))
+        self.config.add_extra("_help",
+                              ns="SOC intervals",
+                              tagc="AGC time intervals in seconds",
+                              socf="Force charge SOC level",
+                              r="ratio of time range 1 to time range 2",
+                              socl="lowest SOC limit",
+                              socu="highest SOC limit",
+                              seed='seed (or None) for random number generator',
+                              )
+        self.config.add_extra("_tex",
+                              ns='N_s',
+                              tagc='T_{agc}',
+                              socf='SOC_f',
+                              r='r',
+                              socl='SOC_{l}',
+                              socu='SOC_{u}',
+                              seed='seed',
+                              )
+        self.config.add_extra("_alt",
+                              ns="int",
+                              tagc="float",
+                              socf="float",
+                              r="float",
[…]
