+++ ams/extension/__init__.py
@@ -1,3 +1,5 @@
 """
 Extension module.
 """
+
+from ams.extension import eva  # NOQA
