### 2024-05-04 815d13f
[WIP] EVA, normal distribution parameter initialization
