### 2024-05-04 49f0e98
[WIP] EVA setup, populate parameters and variables
