+++ ams/extension/eva.py
@@ -2,6 +2,7 @@
 EV Aggregator.
 """
 
+import logging
 from collections import OrderedDict
 
 import scipy.stats as stats
@@ -9,9 +10,12 @@ import scipy.stats as stats
 from andes.core import Config
 from andes.core.param import NumParam
 from andes.core.model import ModelData
-from andes.shared import pd
+from andes.shared import np, pd
+from andes.utils.misc import elapsed
 
 from ams.core.model import Model
+
+logger = logging.getLogger(__name__)
 
 
 class EVA(ModelData, Model):
@@ -30,7 +34,7 @@ class EVA(ModelData, Model):
     """
 
     def __init__(self, N=10000, Ns=20, Tagc=4, SOCf=0.2, r=0.5,
-                 seed=None,):
+                 t=18, seed=None,):
         """
         Initialize the EV aggregation model.
 
@@ -46,69 +50,88 @@ class EVA(ModelData, Model):
             Force charge SOC level between 0 and 1, default is 0.2.
         r : float, optional
             Ratio of time range 1 to time range 2 between 0 and 1, default is 0.5.
+        seed : int or None, optional
+            Seed for random number generator, default is None.
+        t : int, optional
+            Current time in 24 hours, default is 18.
         """
         # inherit attributes and methods from ANDES `ModelData` and AMS `Model`
         ModelData.__init__(self)
         Model.__init__(self, system=None, config=None)
 
+        # internal flags
+        self.is_setup = False        # if EVA has been setup
+
+        self.t = np.array(t, dtype=float)  # time in 24 hours
+
         # manually set config as EVA is not processed by the system
         self.config = Config(self.__class__.__name__)
-        self.config.add(OrderedDict((('ns', Ns),
+        self.config.add(OrderedDict((('n', int(N)),
+                                     ('ns', Ns),
                                      ('tagc', Tagc),
                                      ('socf', SOCf),
                                      ('r', r),
                                      ('socl', 0),
                                      ('socu', 1),
+                                     ('tf', self.t),
+                                     ('prumax', 0),
+                                     ('prdmax', 0),
                                      ('seed', seed),
                                      )))
         self.config.add_extra("_help",
+                              n="Number of related EVs",
                               ns="SOC intervals",
                               tagc="AGC time intervals in seconds",
                               socf="Force charge SOC level",
                               r="ratio of time range 1 to time range 2",
                               socl="lowest SOC limit",
                               socu="highest SOC limit",
+                              tf="EVA running end time in 24 hours",
+                              prumax="maximum power of regulation up, in MW",
+                              prdmax="maximum power of regulation down, in MW",
                               seed='seed (or None) for random number generator',
                               )
         self.config.add_extra("_tex",
+                              n='N_{ev}',
                               ns='N_s',
                               tagc='T_{agc}',
                               socf='SOC_f',
                               r='r',
                               socl='SOC_{l}',
                               socu='SOC_{u}',
+                              tf='T_f',
+                              prumax='P_{ru,max}',
+                              prdmax='P_{rd,max}',
                               seed='seed',
                               )
         self.config.add_extra("_alt",
+                              n='int',
                               ns="int",
                               tagc="float",
                               socf="float",
                               r="float",
                               socl="float",
[…]
