### 2024-05-04 062dfa0
[WIP] EVA setup, populate soc_intv and state
