### 2024-05-05 f236b20
[WIP] EVA setup, adjust SOC after populating
