### 2024-05-05 446a586
[WIP] EVA setup, add state space modeling variables
