### 2024-05-05 048e7d7
[WIP] EVA setup, refactor distribution parameters
