### 2024-05-06 b757ffb
[WIP] EVA setup, refactor as EVD and EVA
