+++ ams/extension/eva.py
@@ -1,5 +1,14 @@
 """
-EV Aggregator.
+EV Aggregator module.
+
+EVD is the generated datasets, and EVA is the aggregator model.
+
+Reference:
+[1] J. Wang et al., "Electric Vehicles Charging Time Constrained Deliverable Provision of Secondary
+Frequency Regulation," in IEEE Transactions on Smart Grid, doi: 10.1109/TSG.2024.3356948.
+[2] M. Wang, Y. Mu, Q. Shi, H. Jia and F. Li, "Electric Vehicle Aggregator Modeling and Control for
+Frequency Regulation Considering Progressive State Recovery," in IEEE Transactions on Smart Grid,
+vol. 11, no. 5, pp. 4176-4189, Sept. 2020, doi: 10.1109/TSG.2020.2981843.
 """
 
 import logging
@@ -37,23 +46,14 @@ udist = {'Pc': {'lb': 5.0, 'ub': 7.0},
          'Q': {'lb': 20.0, 'ub': 30.0}}
 
 
-class EVA(ModelData, Model):
+class EVD(ModelData, Model):
     """
-    State space modeling based EV aggregation model.
-
-    In the EVA, each single EV is recorded as a device with its own parameters.
+    In the EVD, each single EV is recorded as a device with its own parameters.
     The parameters are generated from given statistical distributions.
-
-    Reference:
-    [1] J. Wang et al., "Electric Vehicles Charging Time Constrained Deliverable Provision of Secondary
-    Frequency Regulation," in IEEE Transactions on Smart Grid, doi: 10.1109/TSG.2024.3356948.
-    [2] M. Wang, Y. Mu, Q. Shi, H. Jia and F. Li, "Electric Vehicle Aggregator Modeling and Control for
-    Frequency Regulation Considering Progressive State Recovery," in IEEE Transactions on Smart Grid,
-    vol. 11, no. 5, pp. 4176-4189, Sept. 2020, doi: 10.1109/TSG.2020.2981843.
     """
 
     def __init__(self, N=10000, Ns=20, Tagc=4, SOCf=0.2, r=0.5,
-                 t=18, seed=None, A_csv=None, name='EVA'):
+                 t=18, seed=None, name='EVA', A_csv=None):
         """
         Initialize the EV aggregation model.
 
@@ -73,22 +73,23 @@ class EVA(ModelData, Model):
             Seed for random number generator, default is None.
         t : int, optional
             Current time in 24 hours, default is 18.
-        A_csv : str, optional
-            Path to the CSV file containing the state space matrix A, default is None.
         name : str, optional
             Name of the EVA, default is 'EVA'.
+        A_csv : str, optional
+            Path to the CSV file containing the state space matrix A, default is None.
         """
         # inherit attributes and methods from ANDES `ModelData` and AMS `Model`
         ModelData.__init__(self)
         Model.__init__(self, system=None, config=None)
 
-        # NOTE: Overwrite DataParam `name` to be a string attribute
-        self.name = name
+        self.evdname = name
 
         # internal flags
         self.is_setup = False        # if EVA has been setup
 
         self.t = np.array(t, dtype=float)  # time in 24 hours
+        self.eva = None  # EV Aggregator
+        self.A_csv = A_csv  # path to the A matrix
 
         # manually set config as EVA is not processed by the system
         self.config = Config(self.__class__.__name__)
@@ -150,48 +151,6 @@ class EVA(ModelData, Model):
             for i in range(self.config.ns)
         })
 
-        # states of EV, intersection of charging status and SOC intervals
[…]
