### 2024-05-17 6d950a3
[WIP] EVA setup, docstring
