### 2024-05-17 bf34d1f
Fix bug that PFlow didn't update flag
