### 2024-05-17 c8e2c44
Format, unify PYPOWER routines return variables
