+++ ams/routines/pflow.py
@@ -75,9 +75,8 @@ class PFlow(DCPF):
             raise ValueError(msg)
         ppopt = ppoption(PF_ALG=alg, ENFORCE_Q_LIMS=self.config.qlim)
 
-        res, success, sstats = runpf(casedata=ppc, ppopt=ppopt)
-        self.converged = bool(success)
-        return res, self.converged, sstats
+        res, sstats = runpf(casedata=ppc, ppopt=ppopt)
+        return res, sstats
 
     def run(self, force_init=False, no_code=True, method="newton", **kwargs):
         """
