### 2024-05-17 b9ce40e
Fix mpc reading when gencost is not complete
