### 2024-05-18 4b2a1bd
In DCPF, check results before unpack
