### 2024-05-21 c7efd4a
Skip add gencost when mpc has no gencost
