### 2024-05-22 0cba9ca
Fix OTDF calculation
