### 2024-05-23 eac1fbc
Add tests with dtype=float32
