### 2024-05-23 21ee6da
Add place holder param type into Bus
