+++ docs/source/release-notes.rst
@@ -22,6 +22,7 @@ Frequency Regulation," in IEEE Transactions on Smart Grid, doi: 10.1109/TSG.2024
 
 - Fix OTDF calculation
 - Add parameter `dtype='float64'` in `MatProcessor` PTDF, LODF, and OTDF calculation, to save memory
+- Add placeholder parameter `Bus.type`
 
 v0.9.6 (2024-04-21)
 -------------------
