### 2024-05-24 9769ff0
Refactor MatProcessor tests
