### 2024-06-03 9c2bddb
Typo in MatProcessor docstring
