+++ ams/system.py
@@ -231,14 +231,13 @@ class System(andes_System):
         """
         Set the owner for routine attributes: ``RParam``, ``Var``, and ``RBaseService``.
         """
+        # NOTE: here we skip assigning `MParam` owner as it is assined in `MatProcessor`
         for item_name, item in items.items():
             if item.model in self.groups.keys():
                 item.is_group = True
                 item.owner = self.groups[item.model]
             elif item.model in self.models.keys():
                 item.owner = self.models[item.model]
-            elif item.model == 'mats':
-                item.owner = self.mats
             else:
                 logger.debug(f'item_name: {item_name}')
                 msg = f'Model indicator \'{item.model}\' of <{item.rtn.class_name}.{item_name}>'
