### 2024-06-12 bad7136
Improve build_ptdf to allow partial building
