### 2024-06-12 5fd5535
Use LazyImport for scipy.sparse
