### 2024-06-12 a16b42d
Improve scipy.sparse import in MatProcessor
