### 2024-06-12 13d8004
Improve scipy.sparse import in test_mats
