### 2024-06-12 90795da
Remove parameter no_dense
