### 2024-06-12 11000ec
[WIP] Incrementally build PTDF and LODF
