+++ ams/core/matprocessor.py
@@ -4,13 +4,14 @@ Module for system matrix make.
 
 import logging
 import os
+import sys
 from typing import Optional
 
 import numpy as np
 
-from andes.utils.misc import elapsed
 from andes.thirdparty.npfunc import safe_div
-from andes.shared import pd
+from andes.shared import pd, tqdm, tqdm_nb
+from andes.utils.misc import elapsed, is_notebook
 
 from ams.opt.omodel import Param
 from ams.shared import sps
@@ -143,6 +144,7 @@ class MatProcessor:
     def __init__(self, system):
         self.system = system
         self.initialized = False
+        self.pbar = None
 
         self.Cft = MParam(name='Cft', tex_name=r'C_{ft}',
                           info='Line connectivity matrix',
@@ -431,7 +433,7 @@ class MatProcessor:
         return b
 
     def build_ptdf(self, line=None, dtype='float64', no_store=False,
-                   incremental=False, chunk_size=1000):
+                   incremental=False, chunk_size=1000, no_tqdm=False):
         """
         Build the Power Transfer Distribution Factor (PTDF) matrix and store
         it in the MParam `PTDF` by default.
@@ -457,11 +459,13 @@ class MatProcessor:
         dtype : str, optional
             Data type of the PTDF matrix. Default is 'float64'.
         no_store : bool, optional
-            If True, the PTDF will not be stored into `MatProcessor.PTDF._v`.
+            If False, the PTDF will be stored into `MatProcessor.PTDF._v`.
         incremental : bool, optional
             If True, the sparse PTDF will be calculated in chunks to save memory.
         chunk_size : int, optional
             Chunk size for incremental calculation.
+        no_tqdm : bool, optional
+            If True, the progress bar will be disabled.
 
         Returns
         -------
@@ -508,12 +512,37 @@ class MatProcessor:
         Bf = self.Bf._v
 
         if incremental:
+            # initialize progress bar
+            if is_notebook():
+                self.pbar = tqdm_nb(total=100, unit='%', file=sys.stdout,
+                                    disable=no_tqdm)
+            else:
+                self.pbar = tqdm(total=100, unit='%', ncols=80, ascii=True,
+                                 file=sys.stdout, disable=no_tqdm)
+
+            self.pbar.update(0)
+            last_pc = 0
+
             H = sps.lil_matrix((len(luid), system.Bus.n), dtype=dtype)
 
             for start in range(0, nline, chunk_size):
                 end = min(start + chunk_size, nline)
                 H[start:end, noslack] = sps.linalg.spsolve(Bbus[np.ix_(noslack, noref)].T,
                                                            Bf[np.ix_(luid[start:end], noref)].T).T
+
+                # show progress in percentage
+                perc = np.round(min((end / nline) * 100, 100), 2)
+
+                perc_diff = perc - last_pc
+                if perc_diff >= 1:
+                    self.pbar.update(perc_diff)
+                    last_pc = perc
+
+            # finish progress bar
+            self.pbar.update(100 - last_pc)
+            # removed `pbar` so that System object can be serialized
+            self.pbar.close()
+            self.pbar = None
         else:
             H = np.zeros((nline, nbus), dtype=dtype)
             H[:, noslack] = np.linalg.solve(Bbus.todense()[np.ix_(noslack, noref)].T,
@@ -524,7 +553,8 @@ class MatProcessor:
 
         return H
 
-    def build_lodf(self, dtype='float64', no_store=False):
+    def build_lodf(self, dtype='float64', no_store=False,
[…]
