### 2024-06-12 4e97a26
Add decimals check in PTDF incremental calculation
