+++ ams/core/matprocessor.py
@@ -466,7 +466,8 @@ class MatProcessor:
         no_tqdm : bool, optional
             If True, the progress bar will be disabled.
         decimals : int, optional
-            Number of decimal places to round in the incremental calculation.
+            Number of decimal places to round in the incremental calculation,
+            default is 4.
 
         Returns
         -------
