### 2024-06-16 0402e09
[WIP] Refactor test_mats, PTDF part
