+++ tests/test_mats.py
@@ -99,80 +99,195 @@ class TestMatProcessorBasic(unittest.TestCase):
         np.testing.assert_equal(self.mats.Pbusinj._v.shape, (self.nb,))
 
 
-class TestMatProcessorTDFs(unittest.TestCase):
+# class TestMatProcessorTDFs(unittest.TestCase):
+#     """
+#     Test PTDF, LODF, OTDF.
+#     """
+
+#     def setUp(self) -> None:
+#         self.cases = ['matpower/case14.m',
+#                       'matpower/case39.m',
+#                       'matpower/case118.m']
+
+#     def test_ptdf_before_mat_init(self):
+#         """
+#         Test `build_PTDF()` before MatProcessor initialization `build()`.
+#         """
+
+#         for case in self.cases:
+#             ss = ams.load(ams.get_case(case),
+#                           setup=True, default_config=True, no_output=True)
+
+#             # --- test no_store ---
+#             ptdf = ss.mats.build_ptdf(no_store=True)
+#             self.assertIsNone(ss.mats.PTDF._v)
+#             self.assertEqual(ss.mats.PTDF._v.shape, (ss.Line.n, ss.Bus.n))
+
+#     def test_ptdf_line(self):
+#         """
+#         Test `build_PTDF()` with line.
+#         """
+
+#         for case in self.cases:
+#             ss = ams.load(ams.get_case(case),
+#                           setup=True, default_config=True, no_output=True)
+
+#             flag_build = ss.mats.build()
+#             self.assertTrue(flag_build)
+
+#             # TODO: test different line inputs
+
+#     def test_lodf_before_ptdf(self):
+#         """
+#         Test `LODF` before `PTDF`.
+#         """
+
+#         for case in self.cases:
+#             ss = ams.load(ams.get_case(case),
+#                           setup=True, default_config=True, no_output=True)
+
+#             _ = ss.mats.build_lodf(no_store=True)
+#             self.assertIsNone(ss.mats.LODF._v)
+
+#             _ = ss.mats.build_lodf(no_store=False)
+#             self.assertEqual(ss.mats.LODF._v.shape, (ss.Line.n, ss.Line.n))
+
+#     def test_otdf_before_lodf(self):
+#         """
+#         Test `OTDF`.
+#         """
+
+#         for case in self.cases:
+#             ss = ams.load(ams.get_case(case),
+#                           setup=True, default_config=True, no_output=True)
+#             # build matrices
+#             ss.mats.build()
+
+#             otdf64 = ss.mats.build_otdf()
+#             self.assertEqual(otdf64.shape, (ss.Line.n, ss.Bus.n))
+
+#             # input str
+#             otdf_l2 = ss.mats.build_otdf(line=ss.Line.idx.v[2])
+#             self.assertEqual(otdf_l2.shape, (ss.Line.n, ss.Bus.n))
+
+#             # input list with single element
+#             otdf_l2 = ss.mats.build_otdf(line=ss.Line.idx.v[2:3])
+#             self.assertEqual(otdf_l2.shape, (ss.Line.n, ss.Bus.n))
+
+#             # input list with multiple elements
+#             otdf_l23 = ss.mats.build_otdf(line=ss.Line.idx.v[2:5])
+#             self.assertEqual(otdf_l23.shape, (ss.Line.n, ss.Bus.n))
+
+
+class TestBuildPTDF(unittest.TestCase):
     """
-    Test PTDF, LODF, OTDF.
+    Test build PTDF.
     """
 
     def setUp(self) -> None:
-        self.cases = ['matpower/case14.m',
-                      'matpower/case39.m',
-                      'matpower/case118.m']
+        self.ss = ams.load(ams.get_case('matpower/case14.m'),
[…]
