### 2024-06-16 e0a84d8
[WIP] Refactor test_mats, LODF part in progress
