### 2024-06-17 92fff59
Add incremental building in MatProcessor.build_lodf
