+++ ams/core/matprocessor.py
@@ -433,8 +433,7 @@ class MatProcessor:
         return b
 
     def build_ptdf(self, line=None, no_store=False,
-                   incremental=False, chunk_size=1000, no_tqdm=False,
-                   decimals=4):
+                   incremental=False, chunk_size=1000, no_tqdm=False):
         """
         Build the Power Transfer Distribution Factor (PTDF) matrix and store
         it in the MParam `PTDF` by default.
@@ -465,9 +464,6 @@ class MatProcessor:
             Chunk size for incremental calculation.
         no_tqdm : bool, optional
             If True, the progress bar will be disabled.
-        decimals : int, optional
-            Number of decimal places to round in the incremental calculation,
-            default is 4.
 
         Returns
         -------
@@ -525,13 +521,13 @@ class MatProcessor:
             self.pbar.update(0)
             last_pc = 0
 
-            H = sps.lil_matrix((len(luid), system.Bus.n))
+            H = sps.lil_matrix((nline, system.Bus.n))
 
+            # NOTE: for PTDF, we are building rows by rows
             for start in range(0, nline, chunk_size):
                 end = min(start + chunk_size, nline)
                 sol = sps.linalg.spsolve(Bbus[np.ix_(noslack, noref)].T,
                                          Bf[np.ix_(luid[start:end], noref)].T).T
-                sol = np.round(sol, decimals)
                 H[start:end, noslack] = sol
 
                 # show progress in percentage
@@ -561,7 +557,8 @@ class MatProcessor:
 
         return H
 
-    def build_lodf(self, no_store=False, incremental=False, chunk_size=1000, no_tqdm=False):
+    def build_lodf(self, line=None, no_store=False,
+                   incremental=False, chunk_size=1000, no_tqdm=False):
         """
         Build the Line Outage Distribution Factor matrix and store it in the
         MParam `LODF`.
@@ -577,6 +574,10 @@ class MatProcessor:
 
         Parameters
         ----------
+        line: int, str, list, optional
+            Lines index for which the LODF is calculated. It takes both single
+            or multiple line indices. Note that if `line` is given, the LODF will
+            not be stored in the MParam.
         no_store : bool, optional
             If False, the LODF will be stored into `MatProcessor.LODF._v`.
         incremental : bool, optional
@@ -599,8 +600,23 @@ class MatProcessor:
 
         https://www.powerworld.com/WebHelp/Content/MainDocumentation_HTML/Line_Outage_Distribution_Factors_LODFs.htm
         """
-        nl = self.system.Line.n
+        system = self.system
 
+        if line is None:
+            luid = system.Line.idx2uid(system.Line.idx.v)
+        elif isinstance(line, (int, str)):
+            try:
+                luid = [system.Line.idx2uid(line)]
+            except ValueError:
+                raise ValueError(f"Line {line} not found.")
+        elif isinstance(line, list):
+            luid = system.Line.idx2uid(line)
+
+        # NOTE: here we use nbranch to differentiate it with nline
+        nbranch = system.Line.n
+        nline = len(luid)
+
+        ptdf = self.PTDF._v
         # build PTDF if not built
         if self.PTDF._v is None:
             ptdf = self.build_ptdf(no_store=True, incremental=incremental, chunk_size=chunk_size)
@@ -619,31 +635,34 @@ class MatProcessor:
             self.pbar.update(0)
             last_pc = 0
 
-            H = ptdf @ self.Cft._v
-            h = H.diagonal(0).reshape(1, -1)
[…]
