### 2024-06-17 8411249
Refactor test_mats LODF part
