+++ tests/test_mats.py
@@ -189,7 +189,6 @@ class TestBuildPTDF(unittest.TestCase):
                            setup=True, default_config=True, no_output=True)
         self.nl = self.ss.Line.n
         self.nb = self.ss.Bus.n
-        self.dec = 4
 
         self.assertFalse(self.ss.mats.initialized)
         # NOTE: here we test `no_store=True` option
@@ -213,17 +212,17 @@ class TestBuildPTDF(unittest.TestCase):
         # input str
         ptdf_l2 = self.ss.mats.build_ptdf(line=self.ss.Line.idx.v[2], no_store=False)
         self.assertIsNone(self.ss.mats.PTDF._v)
-        np.testing.assert_array_almost_equal(ptdf_l2, self.ptdf_full[2, :], decimal=self.dec)
+        np.testing.assert_array_almost_equal(ptdf_l2, self.ptdf_full[2, :])
 
         # input list with single element
         ptdf_l2p = self.ss.mats.build_ptdf(line=[self.ss.Line.idx.v[2]], no_store=False)
         self.assertIsNone(self.ss.mats.PTDF._v)
-        np.testing.assert_array_almost_equal(ptdf_l2p, self.ptdf_full[[2], :], decimal=self.dec)
+        np.testing.assert_array_almost_equal(ptdf_l2p, self.ptdf_full[[2], :])
 
         # input list with multiple elements
         ptdf_l23 = self.ss.mats.build_ptdf(line=self.ss.Line.idx.v[2:4], no_store=False)
         self.assertIsNone(self.ss.mats.PTDF._v)
-        np.testing.assert_array_almost_equal(ptdf_l23, self.ptdf_full[2:4, :], decimal=self.dec)
+        np.testing.assert_array_almost_equal(ptdf_l23, self.ptdf_full[2:4, :])
 
     def test_ptdf_incremental_lines(self):
         """
@@ -233,8 +232,7 @@ class TestBuildPTDF(unittest.TestCase):
         ptdf_l2 = self.ss.mats.build_ptdf(line=self.ss.Line.idx.v[2], incremental=True, no_store=False,
                                           no_tqdm=True)
         np.testing.assert_array_almost_equal(ptdf_l2.todense(),
-                                             self.ptdf_full[[2], :],
-                                             decimal=self.dec)
+                                             self.ptdf_full[[2], :])
         self.assertTrue(sps.isspmatrix_lil(ptdf_l2))
         self.assertIsNone(self.ss.mats.PTDF._v)
 
@@ -242,25 +240,14 @@ class TestBuildPTDF(unittest.TestCase):
         ptdf_l2p = self.ss.mats.build_ptdf(line=[self.ss.Line.idx.v[2]], incremental=True, no_store=False,
                                            no_tqdm=False)
         np.testing.assert_array_almost_equal(ptdf_l2p.todense(),
-                                             self.ptdf_full[[2], :],
-                                             decimal=self.dec)
+                                             self.ptdf_full[[2], :])
         self.assertTrue(sps.isspmatrix_lil(ptdf_l2p))
         self.assertIsNone(self.ss.mats.PTDF._v)
 
         # input list with multiple elements
         ptdf_l23 = self.ss.mats.build_ptdf(line=self.ss.Line.idx.v[2:4], incremental=True, no_store=False)
         np.testing.assert_array_almost_equal(ptdf_l23.todense(),
-                                             self.ptdf_full[2:4, :],
-                                             decimal=self.dec)
-
-        # test decimals
-        ptdf_l23_dec2 = self.ss.mats.build_ptdf(line=self.ss.Line.idx.v[2:4], incremental=True,
-                                                no_store=False, decimals=2)
-        self.assertTrue(sps.isspmatrix_lil(ptdf_l23_dec2))
[…]
