### 2024-06-17 4347426
In MatProcessor.build_lodf, use incrementall if PTDF._v is sparse
