### 2024-06-17 d09dedf
Refactor test_mats, OTDF part
