### 2024-06-17 922d7b5
Add test_lodf_incrementa_step
