### 2024-06-18 86a8845
Improve test_known_good
