+++ tests/test_known_good.py
@@ -1,52 +1,264 @@
 import logging
 import unittest
 
+import json
+
 import numpy as np
 
+from andes.shared import rad2deg
 import ams
 
 logger = logging.getLogger(__name__)
 
 
 class TestKnownResults(unittest.TestCase):
-    # NOTE: Baseline comes from MATPOWER v8.0b1
-    # NOTE: when using ECOS in this case, the accuracy is ralatively low
 
-    def test_known_case14(self):
-        case = ams.get_case('matpower/case14.m')
-        sp = ams.load(case, setup=True, no_output=True)
-        sp.DCOPF.run(solver='ECOS')
-        self.assertAlmostEqual(sp.DCOPF.obj.v, 7642.59177715644, places=2)
+    def setUp(self) -> None:
+        with open(ams.get_case('matpower/benchmark.json'), 'r') as file:
+            self.mpres = json.load(file)
 
-        aBus_mp = np.array([
-            -0.0, -0.088451789, -0.22695322, -0.18549695,
-            -0.15942929, -0.25995018, -0.24348912, -0.24348912,
-            -0.27468284, -0.2795552, -0.27334388, -0.27941265,
-            -0.28242718, -0.30074116,])
-        aBus_mp -= aBus_mp[0]
+        self.sp = ams.load(ams.get_case('matpower/case14.m'),
+                           setup=True, no_output=True, default_config=True)
 
-        aBus = sp.DCOPF.aBus.v - sp.DCOPF.aBus.v[0]
-        self.assertTrue(np.allclose(aBus, aBus_mp, atol=1e-4))
+    def test_DCPF_case14(self):
+        """
+        Test DC power flow for case14.
+        """
+        self.sp.DCPF.run()
+        np.testing.assert_allclose(self.sp.DCPF.aBus.v * rad2deg,
+                                   np.array(self.mpres['case14']['DCPF']['aBus']).reshape(-1),
+                                   rtol=1e-2, atol=1e-2)
 
-        pg_mp = np.array([
-            2.2096769, 0.38032305, 0.0, 0.0, 0.0])
-        pg = sp.DCOPF.pg.v
-        self.assertTrue(np.allclose(pg, pg_mp, atol=1e-4))
+        np.testing.assert_allclose(self.sp.DCPF.pg.v * self.sp.config.mva,
+                                   np.array(self.mpres['case14']['DCPF']['pg']).reshape(-1),
+                                   rtol=1e-2, atol=1e-2)
 
-    def test_known_case39(self):
-        case = ams.get_case('matpower/case39.m')
-        sp = ams.load(case, setup=True, no_output=True)
-        sp.DCOPF.run(solver='ECOS')
-        self.assertAlmostEqual(sp.DCOPF.obj.v, 41263.94078592735, places=2)
+    def test_PFlow_case14(self):
+        """
+        Test power flow for case14.
+        """
+        self.sp.PFlow.run()
+        np.testing.assert_allclose(self.sp.PFlow.aBus.v * rad2deg,
+                                   np.array(self.mpres['case14']['PFlow']['aBus']).reshape(-1),
+                                   rtol=1e-2, atol=1e-2)
+        np.testing.assert_allclose(self.sp.PFlow.vBus.v,
+                                   np.array(self.mpres['case14']['PFlow']['vBus']).reshape(-1),
+                                   rtol=1e-2, atol=1e-2)
+        np.testing.assert_allclose(self.sp.PFlow.pg.v.sum() * self.sp.config.mva,
+                                   np.array(self.mpres['case14']['PFlow']['pg']).sum(),
+                                   rtol=1e-2, atol=1e-2)
 
-        pi = sp.DCOPF.pi.v / sp.config.mva
[…]
