### 2024-06-18 e30652f
Change used solver to CLARABEL
