### 2024-06-18 5ebcc3f
Set azure pipline python versions to be 311
