### 2024-06-18 826d539
Set NumPy <= 2.0
