### 2024-06-18 6c30476
In azurepipelines, set kvxopt via conda on Mac and Windows
