### 2024-06-18 bd978b3
Fix azure python versions
