### 2024-07-02 74b2b16
Update cvxpy and numpy version requirements
