+++ README.md
@@ -59,7 +59,9 @@ Use the following resources to get involved.
 
 # Installation
 
-***NOTE:*** ``kvxopt`` is recommended to install via ``conda`` as sometimes ``pip`` struggles to set the correct path for compiled libraries.
+***NOTE:***
+- ``kvxopt`` is recommended to install via ``conda`` as sometimes ``pip`` struggles to set the correct path for compiled libraries
+- `cvxpy` versions **below 1.5** are incompatible with `numpy` versions **2.0 and above**.
 
 AMS is released as ``ltbams`` on PyPI and conda-forge.
 Install from PyPI using pip:
