### 2024-07-02 fafd997
Update release notes
