+++ docs/source/release-notes.rst
@@ -9,6 +9,12 @@ The APIs before v3.0.0 are in beta and may change without prior notice.
 Pre-v1.0.0
 ==========
 
+v0.9.9 (2024-07-01)
+-------------------
+
+- Update dependency requirements to support NumPy v2.0.0
+- Improve the logic of setting `Optz` value
+
 v0.9.8 (2024-06-18)
 -------------------
 
