### 2024-07-02 db5cb91
Fix numpy inf
