+++ ams/shared.py
@@ -16,6 +16,16 @@ from andes.utils.lazyimport import LazyImport
 logger = logging.getLogger(__name__)
 
 sps = LazyImport('import scipy.sparse as sps')
+np = LazyImport('import numpy as np')
+
+# Determine the correct infinity value once and use it throughout
+try:
+    # Try using the newer version attribute first
+    inf = np.inf
+except AttributeError:
+    # Fallback to the older version attribute if the newer one is not available
+    inf = np.Inf
+
 
 # NOTE: copied from CVXPY documentation
 MIP_SOLVERS = ['CBC', 'COPT', 'GLPK_MI', 'CPLEX', 'GUROBI',
