### 2024-07-02 1aea389
Fix numpy nan
