+++ requirements.txt
@@ -1,5 +1,5 @@
 kvxopt>=1.3.2.1
-numpy
+numpy<2.0
 scipy
 sympy>=1.6,!=1.10.0
 pandas
@@ -8,4 +8,4 @@ psutil
 openpyxl
 andes>=1.8.7
 pybind11
-cvxpy>=1.5
+cvxpy
