### 2024-07-02 a397834
Try to fix azure tests error with macOS
