### 2024-07-02 1965653
Try to fix azure tests error with macOS
