+++ azure-pipelines.yml
@@ -40,9 +40,9 @@ jobs:
     vmImage: macOS-latest
   timeoutInMinutes: 360
   strategy:
-    matrix:
-      macos_python310:
-        python.version: '3.10'
+      matrix:
+        Python310:
+          python.version: '3.10'
 
   steps:
   - task: UsePythonVersion@0
