### 2024-07-02 ad85c45
Skip macOS tests on azure pipeline
