### 2024-07-17 #174 @jinningwang
> @jinningwang Still looking for reviewers. Not easy to find people with domain knowledge interested in reviewing software. I found two reviewers who seem to have domain knowledge and experience in python programming. They, however, are looking for 4-6 weeks (rather than 3 weeks). It might make sense as it is summer.
>
> I was wondering if you are okay with that?

Yes, it works for me. Thanks for your efforts, especially considering it is a small crowded domain-software.
