### 2024-07-24 #174 @NimaSarajpoor
[Top comment](https://github.com/pyOpenSci/software-submission/issues/174#issue-2255408589) and  [Editor's Response](https://github.com/pyOpenSci/software-submission/issues/174#issuecomment-2162068899) are updated.
