### 2024-08-19 74b5cea
Follow RTD's deprecation of Sphinx context injection at build time
