+++ docs/source/release-notes.rst
@@ -12,6 +12,8 @@ Pre-v1.0.0
 v0.9.9 (2024-xx-xx)
 -------------------
 
+- Follow RTD's deprecation of Sphinx context injection at build time
+- In MATPOWER conversion, set devices name as None
 - Skip macOS tests in azure-pipelines due to failure in fixing its configuration
 - Prepare to support NumPy v2.0.0, but solvers have unexpected behavior
 - Improve the logic of setting `Optz` value
