### 2024-08-21 e2c4c20
Add parameters for scipy spsolve
