### 2024-08-21 7c5832a
Minor clarification on docstring
