+++ requirements.txt
@@ -1,5 +1,5 @@
 kvxopt>=1.3.2.1
-numpy<3.0
+numpy
 scipy
 sympy>=1.6,!=1.10.0
 pandas
