+++ pyproject.toml
@@ -1,0 +1,66 @@
+[build-system]
+requires = ["setuptools", "wheel", "versioneer[toml]"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "ltbams"
+dynamic = ["version"]
+description = "Python software for scheduling modeling and co-simulation with dynamics."
+readme = "README.md"
+authors = [
+    {name = "Jinning Wang", email = "jinninggm@gmail.com"}
+]
+
+license = {file = "LICENSE"}
+
+classifiers = [
+    "Development Status :: 4 - Beta",
+    "Natural Language :: English",
+    "Programming Language :: Python :: 3",
+    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
+    "Environment :: Console"
+]
+
+dependencies = [
+    "kvxopt>=1.3.2.1",
+    "numpy",
+    "scipy",
+    "sympy>=1.6,!=1.10.0",
+    "pandas",
+    "matplotlib",
+    "psutil",
+    "openpyxl",
+    "andes>=1.8.7",
+    "pybind11",
+    "cvxpy"
+]
+
+[project.optional-dependencies]
+dev = [
+    "pytest",
+    "pytest-cov",
+    "coverage",
+    "flake8",
+    "ipython",
+    "numpydoc",
+    "toml"
+]
+doc = [
+    "sphinx",
+    "pydata-sphinx-theme",
+    "numpydoc",
+    "sphinx-copybutton",
+    "sphinx-panels",
+    "myst-parser",
+    "nbsphinx"
+]
+
+[project.scripts]
+ams = "ams.cli:main"
+
+[tool.versioneer]
+VCS = "git"
+style = "pep440-post"
+versionfile_source = "ams/_version.py"
+versionfile_build = "ams/_version.py"
+tag_prefix = "v"
