### 2024-09-02 ce54651
Minimize setup.py to follow toml
