+++ setup.cfg
@@ -1,6 +1,0 @@
-[versioneer]
-VCS = git
-style = pep440-post
-versionfile_source = ams/_version.py
-versionfile_build = ams/_version.py
-tag_prefix = v
