### 2024-09-02 f61f8d4
Add script to auto generate requirements
