+++ azure-pipelines.yml
@@ -88,6 +88,6 @@ jobs:
       conda install -c conda-forge kvxopt
       python -m pip install --upgrade pip
       pip install pytest pytest-azurepipelines
-      pip install .[dev,interop]
+      pip install .[dev]
       pytest
     displayName: pytest
