+++ .codacy.yml
@@ -3,6 +3,7 @@ engines:
   flake8:
     enabled: true
 exclude_paths:
+  - "pyproject.toml"
   - "setup.py"
   - "versioneer.py"
   - "ams/pypower/**"
