+++ .readthedocs.yml
@@ -14,10 +14,7 @@ sphinx:
 
 python:
   install:
-    - requirements: requirements.txt
     - method: pip
       path: .
       extra_requirements:
-        - doc
-    - method: setuptools
-      path: .
+        - docs
