+++ .readthedocs.yml
@@ -14,7 +14,10 @@ sphinx:
 
 python:
   install:
+    - requirements: requirements.txt
     - method: pip
       path: .
       extra_requirements:
-        - docs
+        - doc
+    - method: setuptools
+      path: .
