### 2024-09-02 fd2cade
Rename genreq to genconf, include versioneer cfg writing
