+++ genconf.py
@@ -1,9 +1,14 @@
 """
-Generate requirements from pyproject.toml.
+Generate configurations from pyproject.toml.
 """
 
 import toml
 from datetime import datetime
+
+
+# Get the current date
+current_date = datetime.now().strftime("%Y-%m-%d")
+comment = f"# Generated on {current_date}.\n"
 
 
 def write_req():
@@ -16,11 +21,6 @@ def write_req():
     dependencies = pyproject['project']['dependencies']
     dev_dependencies = pyproject['project']['optional-dependencies']['dev']
     doc_dependencies = pyproject['project']['optional-dependencies']['doc']
-
-    # Get the current date
-    current_date = datetime.now().strftime("%Y-%m-%d")
-
-    comment = f"# Generated on {current_date}.\n"
 
     # Overwrite requirements.txt
     with open('requirements.txt', 'w') as f:
@@ -40,5 +40,24 @@ def write_req():
     print("Requirements files generated successfully.")
 
 
+def write_cfg():
+    """
+    Write versioneer configuration from pyproject.toml to setup.cfg.
+    """
+    with open('pyproject.toml', 'r') as f:
+        pyproject = toml.load(f)
+
+    versioneer = pyproject['tool']['versioneer']
+
+    with open('setup.cfg', 'w') as f:
+        f.write(comment)
+        f.write("[versioneer]\n")
+        for key, value in versioneer.items():
+            f.write(f"{key} = {value}\n")
+
+    print("Versioneer configuration generated successfully.")
+
+
 if __name__ == "__main__":
     write_req()
+    write_cfg()
