+++ pyproject.toml
@@ -60,6 +60,28 @@ doc = [
 [project.scripts]
 ams = "ams.cli:main"
 
+[project.urls]
+Homepage = "https://ltb.curent.org"
+Documentation = "https://ltb.readthedocs.io/projects/ams/en/latest/getting_started/index.html"
+Repository = "https://github.com/CURENT/ams"
+Issues = "https://github.com/CURENT/ams/issues"
+Changelog = "https://ltb.readthedocs.io/projects/ams/en/latest/release-notes.html"
+
+[tool.setuptools]
+include-package-data = true  # Ensure package data is included
+
+[tool.setuptools.packages.find]
+where = ["ams"]
+
+[tool.setuptools.package-data]
+"*" = ["CONTRIBUTING.rst", "LICENSE", "README.md", "versioneer.py", "ams/_version.py"]
+"cases" = ["*.xlsx", "*.m", "*.raw", "*.dyr", "*.py", "*.json"]
+"docs" = ["*.rst", "*.png", "conf.py", "Makefile", "make.bat"]
+
+[tool.setuptools.exclude-package-data]
+"py" = ["*.pyc", "*.pyo"]
+"pycache" = ["__pycache__"]
+
 [tool.versioneer]
 VCS = "git"
 style = "pep440-post"
