### 2024-09-02 69ceade
Remove MANIFEST.in and setup.cfg
