+++ MANIFEST.in
@@ -1,17 +1,0 @@
-include CONTRIBUTING.rst
-include LICENSE
-include README.md
-include requirements.txt
-include requirements-extra.txt
-
-recursive-exclude * __pycache__
-recursive-exclude * *.py[co]
-
-recursive-include ams/cases *.xlsx *.m *.raw *.dyr *.py *.json
-recursive-include docs *.rst *.png conf.py Makefile make.bat
-
-include versioneer.py
-include ams/_version.py
-
-# If including data files in the package, add them like:
-# include path/to/data_file
