+++ setup.cfg
@@ -1,7 +1,0 @@
-# Generated on 2024-09-02.
-[versioneer]
-VCS = git
-style = pep440-post
-versionfile_source = ams/_version.py
-versionfile_build = ams/_version.py
-tag_prefix = v
