### 2024-09-02 d06e235
Restore setup.cfg
