+++ .github/workflows/pythonapp.yml
@@ -43,6 +43,7 @@ jobs:
     - name: Build a distribution if tagged
       if: github.event_name == 'push' && startsWith(github.ref, 'refs/tags') && matrix.os == 'ubuntu-latest' && matrix.python-version == '3.9'
       run: |
+        pip install build
         python -m build
     - name: Publish a Python distribution to PyPI if tagged
       if: github.event_name == 'push' && startsWith(github.ref, 'refs/tags') && matrix.os == 'ubuntu-latest' && matrix.python-version == '3.9'
