### 2024-09-04 8bc9504
Update release notes
