### 2024-09-04 54681d2
Update release notes
