### 2024-09-04 05e10a2
Update release notes
