+++ tests/test_andes.py
@@ -1,232 +1,0 @@
-"""
-Test ANDES interface.
-"""
-
-import unittest
-import numpy as np
-import pkg_resources
-from pkg_resources import parse_version
-
-import andes
-import ams
-
-from ams.interop.andes import (build_group_table, make_link_table, to_andes,
-                               parse_addfile, verify_pf)
-
-
-class TestMatrices(unittest.TestCase):
-    """
-    Tests for system matrices consistency.
-    """
-
-    andes_version = pkg_resources.get_distribution("andes").version
-    if parse_version(andes_version) < parse_version('1.9.2'):
-        raise unittest.SkipTest("Requires ANDES version >= 1.9.2")
-
-    sp = ams.load(ams.get_case('5bus/pjm5bus_demo.xlsx'),
-                  setup=True, no_output=True, default_config=True,)
-    sa = sp.to_andes(setup=True, no_output=True, default_config=True,)
-
-    def setUp(self) -> None:
-        """
-        Test setup.
-        """
-
-    def test_build_y(self):
-        """
-        Test build_y consistency.
-        """
-        ysp = self.sp.Line.build_y()
-        ysa = self.sa.Line.build_y()
-        np.testing.assert_equal(np.array(ysp.V), np.array(ysa.V))
-
-    def test_build_Bp(self):
-        """
-        Test build_Bp consistency.
-        """
-        Bp_sp = self.sp.Line.build_Bp()
-        Bp_sa = self.sa.Line.build_Bp()
-        np.testing.assert_equal(np.array(Bp_sp.V), np.array(Bp_sa.V))
-
-    def test_build_Bpp(self):
-        """
-        Test build_Bpp consistency.
-        """
-        Bpp_sp = self.sp.Line.build_Bpp()
-        Bpp_sa = self.sa.Line.build_Bpp()
-        np.testing.assert_equal(np.array(Bpp_sp.V), np.array(Bpp_sa.V))
-
-    def test_build_Bdc(self):
-        """
-        Test build_Bdc consistency.
-        """
-        Bdc_sp = self.sp.Line.build_Bdc()
-        Bdc_sa = self.sa.Line.build_Bdc()
-        np.testing.assert_equal(np.array(Bdc_sp.V), np.array(Bdc_sa.V))
-
-
-class TestInteropBase(unittest.TestCase):
-    """
-    Tests for basic function of ANDES interface.
-    """
-    ad_cases = [
-        'ieee14/ieee14_full.xlsx',
-        'ieee39/ieee39_full.xlsx',
-        'npcc/npcc.xlsx',
-    ]
-    am_cases = [
-        'ieee14/ieee14.json',
-        'ieee39/ieee39.xlsx',
-        'npcc/npcc_uced.xlsx',
-    ]
-
-    def setUp(self) -> None:
-        """
-        Test setup.
-        """
-
-    def test_basic_functions(self):
-        """
-        Test basic functions defined in ANDES Interop.
-        """
-        for ad_case in self.ad_cases:
-            sa = andes.load(andes.get_case(ad_case),
-                            setup=True, no_output=True, default_config=True,)
-            # --- test build_group_table ---
-            ssa_stg = build_group_table(adsys=sa,
-                                        grp_name='StaticGen',
-                                        param_name=['u', 'name', 'idx', 'bus'],
-                                        mdl_name=[])
-            self.assertEqual(ssa_stg.shape,
-                             (sa.StaticGen.n, 4))
-
-            ssa_gov = build_group_table(adsys=sa, grp_name='TurbineGov',
-                                        param_name=['idx', 'syn'],
-                                        mdl_name=[])
-            self.assertEqual(ssa_gov.shape,
[…]
