### 2024-09-25 14bbe5e
Minor refactor on pflow_dict
