### 2024-09-25 02dbb1f
Move helper function create_entry into module utils
