### 2024-09-25 7d4badb
Avoid using pandas fillna
