### 2024-09-25 b483674
Fix deprecation related to pkg_resources
