### 2024-09-25 cacd1a0
Specify solver in test_dctypes
