### 2024-09-25 3954245
Update release notes
