### 2024-10-06 #174 @NimaSarajpoor
Here is the review provided by one group of the reviewers, @Alizak-Mech @AlirezaAlamgir, that collaborated together. The reviewers (one mechanical engineer and one energy engineer) wanted to use AMS to simulate their case study that involves a smart [power distribution] grid with several Electrical Vehicle (EV) stations, where each station has its own energy storage system. To review, the reviewers decided to use AMS for their case study. I assume the reviewers have the mathematical form of the objective function, and the constraints, and wanted to use the AMS library to find the optimal solution for their problem.

**The following challenges were reported:**

(1) The authors did a good job in providing examples on how this library accepts different input / output formats. However, some technical details seem to be unclear, which makes it hard for [energy / mechanical] engineers who wants to use this library.

(2) The abbreviations seem to be a bit unclear and it is not easy to understand. If users need to know the  naming conventions in this library, it might be a good idea to have a table to explain them, and maybe add a link to that in the README page. It is recommended, if possible, to improve naming for the sake of readability and ease of use.

(3) The reviewers cannot understand how they should define their objective function in AMS.

(4) To motivate users to use this library, the advantages need to be indicated.
* Is it simpler?
* Is it faster?
* Is it more accurate?
* etc?

@Alizak-Mech @AlirezaAlamgir
Feel free to add/correct if needed.

---

Note1: My own review, using [the review template](https://www.pyopensci.org/software-peer-review/appendices/templates.html#peer-review-template) with some extra notes, will be provided in my next comment.

Note2: Since the accuracy and running time depend (highly) on the "optimization solver", this editor believes the authors of library has focused on the "simplicity" by abstracting away some complex layers one may face in building a case study using other libraries. [More on this in my own, upcoming, comment]
