### 2024-10-07 #174 @NimaSarajpoor
**To fill the gap, I decided to do a review myself. Kindly find my review below. If you notice something is off, please let me know. Before that, I first want to acknowledge the fact that the case studies in power system can become very complicated, and the author(s) of library did a great job in creating a software that can handle different kinds of studies (referred to as "routines" in the library IIUC).**

**The package review is provided below. The parts `extra notes` are my additional comments..**

Please check off boxes as applicable, and elaborate in comments below. Your review is not limited to these topics, as described in the reviewer guide*

**Extra notes:**
- I can see a table and figure in the README that try to provide some evidence. The information there is to show two things: (1) The result is accurate, and (2) the running time is reasonable. These two items highly depend on the optimization solver. So, I believe the purpose of the evidence was to show that it is comparable to other libraries. But, the "advantage" of using this library needs to be stated as well. If the goal of library is to provide a "simple" interface for users to handle power system case studies, then README should try to bring user's attention to it.

- The figure is a bit confusing. The legend shows one colour for MATPOWER but several colours for AMS. User cannot easily understand at what they should look there. It is better to remove unnecessary information. For instance, is the running time for `AMS init` necessary? IIUC, `AMS init` gives initial answer which is not optimal. So, is there any point in reporting that?

### Readme file  requirements

    - [x] License

**Extra notes:**
* I can see citation in API document. Maybe create a section in README as well (maybe just add a link to that)

### Usability

**Extra note:**
* There is [this nice table](https://ltb.readthedocs.io/projects/ams/en/latest/routineref.html) that contains different types of routines. It might be a good idea to provide a corresponding example per each routine.

* The API documentation page does not show `AMS 0.9.10` properly on top left of the page.

* Regarding `AMS xlsx` input format, I noticed that a link is provided that takes user to a page that contains information about `ANDES xlsx`. I understand that the expectation from user is to click on that and go through the doc. Still, I think some common names can be defined in AMS too. For instance, maybe just copy the definition of common parameters and provide it in AMS API doc as well.

* On API doc, I started from installation and click on next... till I get to [the example simulate](https://ltb.readthedocs.io/projects/ams/en/latest/_examples/ex1.html). However, there are some gaps that should have been filled first. For instance, the example talks about Device, Model, and Routine. For instance, it says: "model refers to the device-level models".  However, the definition of `device`, and `device-level model` themselves are not clear.

* Each routine is a particular study in power system that tries to find an optimal solution for an optimization problem. In my opinion, one of the most important parts of API doc is [this table of routines](https://ltb.readthedocs.io/projects/ams/en/latest/routineref.html). Because, each routine is linked to a page that clearly shows params, variables, constraints, etc. One thing that can improve it further is to add a link to an example notebook as well. Sharing that link in the README might be helpful for users.

* In API referemce, the first item is `ams.system.System`, and the first argument is "case". However, the definition of "case" may not be clear for all. I understand that it is an input path to the data of a case study but I believe it should have been mentioned in the description of the argument.

* Again, in `ams.system`, I noticed the description of `system.add` says: "Add a device instance for an existing model." However, I believe the definition of device / model has not yet been revealed to the user by that point. So, in the very beginning of the document, letting user know the common terminologies and their definition can be very helpful.

- [?] **Functionality:** Any functional claims of the software been confirmed.
- [N/A] **Performance:** Any performance claims of the software been confirmed.

**Extra note:**
* Re: "Any functional claims of the software been confirmed": Too complicated to be checked. But, I ran the examples and the classes seem to be working as expected.

* Re: "Any performance claims of the software been confirmed.": By performance, if it means accuracy, I can say it is confirmed by running the examples myself. If it means the running time, then I can't confirm it since our hardwares do not match.
