### 2024-10-07 #174 @NimaSarajpoor
@lwasser
Maybe help find a second person please let me know? Sure! That should be very helpful!

**[Just a suggestion]**
According to my experience with this library, I think the focus of the new person can be more on the documentation part.
