### 2024-11-03 #174 @Kawians
**Peer Review Notes**
- **Problem Definition:** While it mentions AMS's functionality (stability-constrained modeling and credible scheduling), the README lacks a clear, concise statement on the specific problems AMS solves. For example, defining challenges in power system stability, cost optimization, or integration with existing simulation engines would clarify the exact problems it addresses.
- **Target Audience:** Although implied, the README does not directly specify the target audience (e.g., researchers in power systems, energy market analysts, or grid operators). Explicitly stating who would benefit from AMS could strengthen this section.

**Peer Review Note:**
- Repostatus badge could be added to the readme file.

**Peer Review Notes:**
* There is no demonstration of package usage in the README. Adding a simple code snippet or example would make it clearer for users.
* The Citation to your work could be found on your Documentation web page. Just as a suggestion, it could be added to your Readme file if needed as well.

**Peer Review Note:**
- The documentation is well-structured, featuring a clear table of contents that includes sections such as "Getting started," "Examples," "Development," "Release notes," "Routine reference," "Model reference," and "API reference." This organization facilitates easy navigation and access to relevant information. The "Getting Started" section provides concise installation instructions and links to further resources, enhancing user accessibility.

**Peer Review Note:**

- Installation Challenge:
I had difficulties installing the package using pip install on the Mac intel series in Late 2015. It took forever, caused the laptop to get very hot, and frozen, and the package never got installed. However, the conda install worked properly (but again for a bit of a long time about 40 minutes if I remember right).
To make sure it is not my computer's problem I tried it on a newer version of the Mac intel series and it worked properly. However it seems it could be only my computer issue, I decided to bring this up with you to further investigate it if needed.

- Functionality and Performance:
I second what Nima mentioned. The examples work as expected, but it is hard to generalize the functionality and performance.
