### 2024-12-01 #174 @NimaSarajpoor
@Alizak-Mech @AlirezaAlamgir
I understand that your schedule might be tight. If that is the case, can you please let us know if there is a work-in-progress regarding template? Or, should we rely on the feedback provided to me before and shared here https://github.com/pyOpenSci/software-submission/issues/174#issuecomment-2395477842?
