### 2025-08-24 #174 @jinningwang
@ab93 @lwasser Hi, my response is as below, and for easier reading, a PDF copy is also attached here: [AMS-submission.pdf](https://github.com/user-attachments/files/21959455/AMS-submission.pdf)

Hi there, thank you all for the comments and efforts on this submission!

# Summary

The package version has been updated to v1.0.13, compared to the version submitted v0.9.6.
Here is a summary we've made given the feedback from here:

1. A statement of target users is added to README Why AMS section, <https://github.com/CURENT/ams/blob/develop/README.md#why-ams>
1. A code snippet demonstrate the usage of the package is added to README Example Usage section, <https://github.com/CURENT/ams/blob/develop/README.md#example-usage>
1. Some brief comparasion with other tools is added to README Why AMS section, <https://github.com/CURENT/ams/blob/develop/README.md#why-ams>
1. Citation information is add to the README Citing AMS section, <https://github.com/CURENT/ams/blob/develop/README.md#citing-ams>
1. A repostatus.org badge is added in the README badges table
1. Extra notes on installation are added to the README Installation section, <https://github.com/CURENT/ams/blob/develop/README.md#installation>

Some changes are made to the develop branch and will be merged into the master branch later.

# Comments

## Reviewer Alizak-Mech and AlirezaAlamgir,

> Point: The authors did a good job in providing examples on how this library accepts different input / output formats. However, some technical details seem to be unclear, which makes it hard for [energy / mechanical] engineers who wants to use this library.

Response: In the project documentation, we add a summary table of some common acronyms used in the library, <https://ltb.readthedocs.io/projects/ams/en/develop/routineref.html#routine-reference>

> Point: The abbreviations seem to be a bit unclear and it is not easy to understand. If users need to know the naming conventions in this library, it might be a good idea to have a table to explain them, and maybe add a link to that in the README page. It is recommended, if possible, to improve naming for the sake of readability and ease of use.

Response: We add a link to the acronym table in the README Example Usage section, <https://github.com/CURENT/ams/blob/develop/README.md#example-usage>

> Point: The reviewers cannot understand how they should define their objective function in AMS.

Response: We add more details in the documentation to clarify how to define an objective function in AMS, <https://ltb.readthedocs.io/projects/ams/en/stable/modeling/example.html#model-section>

> Point: To motivate users to use this library, the advantages need to be indicated.
> - Is it simpler?
> - Is it faster?
> - Is it more accurate?
> - etc?

Response: In README, we provide:
1. A brief code snippet showing the usage, which shows simpler usage.
1. A table of results comparasion with other libraries, which shows accuracy.
1. Two figures of computation time, which shows speed.

More in-depth discussion can be found in the [design philosophy paper][paper] Case Study section.

> Point: Since the accuracy and running time depend (highly) on the "optimization solver", this editor believes the authors of library has focused on the "simplicity" by abstracting away some complex layers one may face in building a case study using other libraries. [More on this in my own, upcoming, comment]

Response: In the README Why AMS section, we add the second paragraph to further clarify this: <https://github.com/CURENT/ams/blob/develop/README.md#why-ams>

> Point: I can see a table and figure in the README that try to provide some evidence. The information there is to show two things: (1) The result is accurate, and (2) the running time is reasonable. These two items highly depend on the optimization solver. So, I believe the purpose of the evidence was to show that it is comparable to other libraries. But, the "advantage" of using this library needs to be stated as well. If the goal of library is to provide a "simple" interface for users to handle power system case studies, then README should try to bring user's attention to it.

Response: In the README Why AMS section, we highlight the major advantages as "AMS facilitates Dynamics Incorporated Scheduling and Scheduling-Dynamics Co-Simulation through an integrated interface with ANDES.": <https://github.com/CURENT/ams/blob/develop/README.md#why-ams>

> Point: The figure is a bit confusing. The legend shows one colour for MATPOWER but several colours for AMS. User cannot easily understand at what they should look there. It is better to remove unnecessary information. For instance, is the running time for AMS init necessary? IIUC, AMS init gives initial answer which is not optimal. So, is there any point in reporting that?

Response: In the README Why AMS section, we add more elaborations on the two figures. The init time is necessary to demonstrate the software scalability. We can see that with the problem size increasing, the init time increases very slowly compared to the overall computation time. More discussion can be found in the [design philosophy paper][paper] Case Study section.

## Reviewer NimaSarajpoor

> Point: Re: "Any functional claims of the software been confirmed": Too complicated to be checked. But, I ran the examples and the classes seem to be working as expected.

Response: Thanks for the confirmation. We acknowledge that the complicated functions in this software, and we have tests on each single routine to ensure they work as intended, such as ``test_rtn_acopf`` and ``test_rtn_rted``, <https://github.com/CURENT/ams/tree/master/tests>

> Point: Re: "Any performance claims of the software been confirmed.": By performance, if it means accuracy, I can say it is confirmed by running the examples myself. If it means the running time, then I can't confirm it since our hardwares do not match.

Response: The running time part can be found in the [design philosophy paper][paper] Case Study section.

## Reviewer Kawians

> Point: Installation Challenge:
> I had difficulties installing the package using pip install on the Mac intel series in Late 2015. It took forever, caused the laptop to get very hot, and frozen, and the package never got installed. However, the conda install worked properly (but again for a bit of a long time about 40 minutes if I remember right). To make sure it is not my computer's problem I tried it on a newer version of the Mac intel series and it worked properly. However it seems it could be only my computer issue, I decided to bring this up with you to further investigate it if needed.

Response: Thanks for bringing this out. We add more installation notes to the README Installation section, <https://github.com/CURENT/ams/blob/develop/README.md#installation>

> Point: Functionality and Performance:
> I second what Nima mentioned. The examples work as expected, but it is hard to generalize the functionality and performance.

Response: Thanks for the confirmation. We acknowledge that the complicated functions in this software, and we have tests on each single routine to ensure they work as intended, such as ``test_rtn_acopf`` and ``test_rtn_rted``, <https://github.com/CURENT/ams/tree/master/tests>

---

* * *

[paper]:                 https://ning.guru/assets/pdf/10836855.pdf
