### 2025-09-08 #174 @ab93
---

🎉 AMS has been approved by pyOpenSci! Thank you @jinningwang for submitting AMS and many thanks to @NimaSarajpoor, @Alizak-Mech, @AlirezaAlamgir and @Kawians for reviewing this package!!

- [x] Add the badge for pyOpenSci peer-review to the README.md of <package-name-here>. The badge should be `[![pyOpenSci Peer-Reviewed](https://pyopensci.org/badges/peer-reviewed.svg)](https://github.com/pyOpenSci/software-review/issues/issue-number)`.

---
