### 2025-11-08 #174 @jinningwang
@ab93 I've improved the repo README, https://github.com/CURENT/ams

Let me know if there is any further concerns.
