### 2025-11-12 #174 @lwasser
it looks like we are missing a zenodo doi in the header and there a few of the check boxes need to be completed in the final editor checks!
